intial code

+++ cards/src/main/java/com/javaexpress/cards/repository/CardsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.javaexpress.cards.model.Cards;
 
 @Repository
-public interface CardsRepository extends CrudRepository<Cards, Integer> {
+public interface CardsRepository extends CrudRepository<Cards, Long> {
 
 	List<Cards> findByCustomerId(int customerId);
 }
